// src/modules/pre-market/pre-market-neighborhood-lookup.generated.ts
export interface PreMarketNeighborhoodLookupEntry {
  neighborhood: string;
  region: string;
  borough: string;
}